fix: regression caused by 35d955c10bd97150a41c2cd748f2c6deb94bb5b8
the mentioned commit fixes #84, a bug where gateway targets could becoming listening to other targets events
if they share the same ID. There was an error made in the migration for this change, where
CHANNEL_CREATE and GUILD_CREATE events were being treated as listens for user objects, not the respective channel or guild
objects. The confusion was due to the return type of splitQualifiedMention containing a 'user' key.